requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
websockets==14.1
httpx==0.28.1
openai==1.58.1
pydantic==2.10.4
pydantic-settings==2.7.1
grpcio==1.69.0
protobuf==5.29.3
python-dotenv==1.0.1
pytest==8.3.4
pytest-asyncio==0.24.0
